refactor request
